Move decoder failures under `Decoder.Failure`
